+++ src/core/se/jbee/inject/event/EventReflector.java
@@ -6,10 +6,10 @@ package se.jbee.inject.event;
  * This should allow users to use their own mechanism and logic. For example the
  * use of user annotations to describe isolation and such properties that are
  * check by the users implementation of this interface to convert them to a
- * universal description {@link EventProperties}.
+ * universal description {@link EventPreferences}.
  */
 @FunctionalInterface
 public interface EventReflector {
 	
-	EventProperties reflect(Class<?> event);
+	EventPreferences reflect(Class<?> event);
 }