+++ src/test/se/jbee/inject/SuiteSilk.java
@@ -12,6 +12,7 @@ import se.jbee.inject.event.SuiteEvent;
 @RunWith ( Suite.class )
 @SuiteClasses ( { TestName.class, TestType.class, TestPackages.class, TestMoreApplicable.class,
 		TestTarget.class, TestDeclarationType.class, TestEqualsHashCode.class,
+		TestSerializable.class,
 		// suits
 		SuitContainer.class, SuiteBind.class, SuiteAction.class, SuiteEvent.class } )
 public class SuiteSilk {